fix some bugs

+++ app/src/main/java/com/hugboga/custom/data/bean/BarginBean.java
@@ -16,7 +16,7 @@ public class BarginBean implements IBaseBean{
     public String userAvatar;       //用户头像
     public String userName;         //用户真实 名称
     public String cnstr;            //服务文案
-    public int multiple;            //倍数
+    public double multiple;         //倍数
     public int isStart;             //砍价是否开始0：未开始；1开始；2已结束
     public List<BarginWebchatList> bargainWechatRspList;
